# perf(gen): run generation in parallel